Update cpu.sh
Changed some functionality in cpu::freq_check -- probably needs a once over to see why this type of logic is needed between CLI vs UI and to unify it back to a single piece of logic.  I am guessing issues with 'array::contains' function but couldn't find a work around.