Update to manage_dtoverlays.sh for overlay_prefix handling (revised) (#615)
* Update manage_dtoverlays.sh, overlay_prefix handling

For issues #612 and #592 where overlays don't load because the overlay_prefix gets written to armbianEnv.txt along with the overlay name. See #612 and #592 for details.

* Update manage_dtoverlays.sh per rabbit review comments

Updated per code rabbit suggestions and added handling the presence of a hyphen "-" following the prefix as a separate check.

* Update manage_dtoverlays.sh

Fixed issue with overlay status update failing due to remaining "-" not being removed along with overlay_prefix.