fix(jellyfin): expose MPP/RGA, setup permissions, rectify purge script (#661)
* fix(hwacc): use correct flag for RKMPP detection

The prior flag `"${LINUXFAMILY}" == "rockchip64"` was returning `false` as the actual value was set to `rk35xx` (tested on ROCK 5T, Noble 25.8.1 Minimal), resulting in no access to hardware acceleration from within the container.

Wasn't affecting direct playback, but using any kind of transcoding were resulting in playback error in the Jellyfin clients, and `ffmpeg` error 187 in the container logs.

* fix(hwacc): add udev rules for RKMPP

In addition to activating the RK35XX conditional block to expose the VPU/MPP hardware, we also need to add the correct `udev` rules (to fix up the group permissions) so that the `jellyfin`  container (rather, `jellyfin-ffmpeg` within the container) can access the host's HWA features (MPP and RGA).

* chore(formatting): remove stray indentation spaces, use EOT instead of EOF

* fix(purge): use correct app installation path for purging

* fix(purge): only delete the config, to avoid deleting user media by accident

* Fix EOT definitions and style - add tabs

* Drop udev rules upon app removal

---------

Co-authored-by: Igor Pecovnik <igor@armbian.com>